refactor(server): split undeliverable-bounce code out of oversized modules
Review fixes — extract the #1444 undeliverable-negotiation constructs
into behavior-grouped modules, no logic changes:

- handlers/jingle/undeliverable.rs: rollback derivation +
  credential-free echo (and their unit tests) move out of the 2.6k-line
  jingle.rs; public paths preserved via re-exports.
- tests/xep0166_jingle/ becomes a directory target: main.rs keeps the
  core suite, undeliverable.rs holds the bounce rollback/echo tests.
- interpret/tests/undeliverable_bounce.rs: the offline-bounce cluster
  (fixtures + 4 route_to_connection tests) moves out of the 6.5k-line
  interpret tests module.
